+++ DIFI_Validator/utils/difi_constants.py
